_scripts/pyproject.toml
[project]
name = "e17-beer-scripts"
dynamic = ["version"]
requires-python = ">=3.11"
dependencies = [
    "google-maps-places>=0.8",
    "pandas>=3",
    "python-dotenv>=1.2",
]

[dependency-groups]
dev = [
    "click>=8",
    "mypy>=1.20",
    "pandas-stubs>=3",
    "pytest>=9",
    "pytest-mock>=3.15",
]

[tool.setuptools]
packages = []  # disable automatic discovery

[tool.ruff]
line-length = 120

[tool.mypy]
python_executable = ".venv/bin/python3"
